CRAFTERCMS-368: Supporting the new design. It still requires more testing

+++ server/src/main/java/org/craftercms/social/services/CounterService.java
@@ -2,6 +2,6 @@ package org.craftercms.social.services;
 
 public interface CounterService {
 	
-	int getNextSequence(String collectionName);
+	long getNextSequence(String collectionName);
 
 }
